+++ embedded-graphics/src/primitives/ellipse/styled.rs
@@ -89,10 +89,12 @@ where
     }
 }
 
-impl<C> Drawable<C> for Styled<Ellipse, PrimitiveStyle<C>>
+impl<C> Drawable for Styled<Ellipse, PrimitiveStyle<C>>
 where
     C: PixelColor,
 {
+    type Color = C;
+
     fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
     where
         D: DrawTarget<Color = C>,