+++ embedded-graphics/src/primitives/rectangle/styled.rs
@@ -82,10 +82,12 @@ where
     }
 }
 
-impl<C> Drawable<C> for Styled<Rectangle, PrimitiveStyle<C>>
+impl<C> Drawable for Styled<Rectangle, PrimitiveStyle<C>>
 where
     C: PixelColor,
 {
+    type Color = C;
+
     fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
     where
         D: DrawTarget<Color = C>,