+++ embedded-graphics/src/primitives/triangle/styled.rs
@@ -76,10 +76,12 @@ where
     }
 }
 
-impl<C> Drawable<C> for Styled<Triangle, PrimitiveStyle<C>>
+impl<C> Drawable for Styled<Triangle, PrimitiveStyle<C>>
 where
     C: PixelColor,
 {
+    type Color = C;
+
     fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
     where
         D: DrawTarget<Color = C>,